refactor(preview-discovery): rename package to ee.schimke.composeai.discovery
`PreviewData.kt` and its test live in the `:preview-discovery` module
since #1300 — but kept the `ee.schimke.composeai.plugin` package to
defer the import churn. Phase A2a closes that gap: package matches the
module name, and the import seam between the gradle plugin and the
new pure-JVM library is now visible at the source level.

24 importers across `:gradle-plugin` pick up `import
ee.schimke.composeai.discovery.*` — star-form because the discovery
types are a tightly-coupled domain (the previews.json schema) and
listing 30+ explicit imports per file makes the diff harder to read
than it solves. Future scan-logic extraction (A2b) lands in the same
package alongside the data types.
