+++ gradle-plugin/preview-discovery/src/main/kotlin/ee/schimke/composeai/discovery/PreviewData.kt
@@ -1,4 +1,4 @@
-package ee.schimke.composeai.plugin
+package ee.schimke.composeai.discovery
 
 import kotlinx.serialization.Serializable
