+++ gradle-plugin/src/main/kotlin/ee/schimke/composeai/plugin/DiscoverAndroidResourcesTask.kt
@@ -1,5 +1,6 @@
 package ee.schimke.composeai.plugin
 
+import ee.schimke.composeai.discovery.*
 import java.io.File
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
