+++ gradle-plugin/src/main/kotlin/ee/schimke/composeai/plugin/ResourceDiscovery.kt
@@ -1,5 +1,6 @@
 package ee.schimke.composeai.plugin
 
+import ee.schimke.composeai.discovery.*
 import java.io.File
 
 /**
