+++ gradle-plugin/src/test/kotlin/ee/schimke/composeai/plugin/ResourceDiscoveryTest.kt
@@ -1,6 +1,7 @@
 package ee.schimke.composeai.plugin
 
 import com.google.common.truth.Truth.assertThat
+import ee.schimke.composeai.discovery.*
 import java.io.File
 import org.junit.Rule
 import org.junit.Test
